refactor(dav): extract /dav/* to built-in plugin
